requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=0.21.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
PyMySQL>=1.0
